+++ frontend/src/metabase/metadata/pages/DataModel/components/TablePicker/TablePicker.tsx
@@ -7,7 +7,7 @@ import EmptyState from "metabase/components/EmptyState";
 import { useDispatch } from "metabase/lib/redux";
 import { Box, Icon, Input, Stack } from "metabase/ui";
 
-import { Results } from "./Item";
+import { Results } from "./Results";
 import type { TreePath } from "./types";
 import {
   flatten,
